Huge style update
This is a pretty big update, that touches four files and adds two new
ones.

This is an all in one solution, because it keeps the archive in a usable
state. There are NO CHANGES to the code at all; everything that changed
is done to clarify situations and improve the overall style of the code.

The test suite changed a bit. These changes still need to be documented.

We basically got multiple test-* targets. 'make test' runs the part of
the test suite, that should *always* succeed. 'make test-all' runs
absolutely everything. It should generally work at any time, too; but
it is mainly intended to check every possible corner of the code before
doing releases.

(This commit contains a small documentation update, too.)