Switch git Getopt::Long
About time. Arename has a shitload of options, I ran out of useful single
letters a long time ago. So this is a sensible thing to do.

This does break backwards compatibility, hence a major version update
should be done. Apart from that, that is as much as I care.