Fixed a test in tests/0700_templates.t
The last test did this:

    expand: &{bar:\&{album?&album!album}

and with `bar' being an unset tag:

    expect: &{album?Red!album}

Which is clearly wrong. The "&{" and the "}" are a matched pair,
enclosing the complex expression inside. Therefore, they're taken.

What remains is this: "\&{album?&album!album". The `&' is quoted, so
it's scanned as a literal ampersand which leaves the `{' alone. Alone it
doesn't have any special meaning, so it gets inserted literally, too.
The same goes for the rest of the string. Nothing in there has special
meaning, so it's read literally, which makes "&{album?Red!album" the
correct result.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>