Fix tracknumber handling
Audio::Scan returns "x/y" type tracknumbers untampered. We used to just
use the `x' part of it. This brings us back in line with how arename3
used to work. The `y' part (if it exists) will be available as a new tag
called `tracknummax'.

Also, the `tnpad' setting wasn't applied. This should fix it. A simple
Test-case for default tnpad-behaviour is included.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>