Add `filetype_known' hook
This hook is run as soon as arename knows whether the current file is a
supported file format. It is run with file name, extension and
type-identifier as arguments.

If the `usetypeasextension' option is set to `false', this hook may be
used to alter the extension which will be used for generating the final
file name.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>