Add support for mp4 files
Audio::Scan has support for it. Adding support for Audio::Scan-enabled
file types should be as simple as this patch. Some may have to do a
little more hacking to get tags in line (like the mp3 integration),
but basically, everything to add support for more file types is already
there.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>